Move addition to collection to general consideration step
