+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/workflows/generalreferral/GeneralReferralWorkflow.java
@@ -9,8 +9,6 @@ import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.DefaultWorkf
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.WorkflowException;
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.Task;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralApplicationAddedDateTask;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralReferralDataTask;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralReferralFieldsRemovalTask;
 
 import java.util.Map;
 
@@ -22,8 +20,6 @@ import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.Orchestrati
 public class GeneralReferralWorkflow extends DefaultWorkflow<Map<String, Object>> {
 
     private final GeneralApplicationAddedDateTask generalApplicationAddedDateTask;
-    private final GeneralReferralDataTask generalReferralDataTask;
-    private final GeneralReferralFieldsRemovalTask generalReferralFieldsRemovalTask;
 
     public Map<String, Object> run(CaseDetails caseDetails) throws WorkflowException {
 
@@ -33,9 +29,7 @@ public class GeneralReferralWorkflow extends DefaultWorkflow<Map<String, Object>
 
         return this.execute(
             new Task[] {
-                generalApplicationAddedDateTask,
-                generalReferralDataTask,
-                generalReferralFieldsRemovalTask
+                generalApplicationAddedDateTask
             },
             caseDetails.getCaseData(),
             ImmutablePair.of(CASE_ID_JSON_KEY, caseId)
