+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/workflows/generalreferral/GeneralConsiderationWorkflowTest.java
@@ -7,13 +7,15 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
+import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralReferralDataTask;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralReferralDecisionDateTask;
+import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralReferralFieldsRemovalTask;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import static uk.gov.hmcts.reform.divorce.orchestration.testutil.Verificators.mockTasksExecution;
-import static uk.gov.hmcts.reform.divorce.orchestration.testutil.Verificators.verifyTaskWasCalled;
+import static uk.gov.hmcts.reform.divorce.orchestration.testutil.Verificators.verifyTasksCalledInOrder;
 
 @RunWith(MockitoJUnitRunner.class)
 public class GeneralConsiderationWorkflowTest extends TestCase {
@@ -21,16 +23,30 @@ public class GeneralConsiderationWorkflowTest extends TestCase {
     @Mock
     private GeneralReferralDecisionDateTask generalReferralDecisionDateTask;
 
+    @Mock
+    private GeneralReferralDataTask generalReferralDataTask;
+
+    @Mock
+    private GeneralReferralFieldsRemovalTask generalReferralFieldsRemovalTask;
+
     @InjectMocks
     private GeneralConsiderationWorkflow generalConsiderationWorkflow;
 
     @Test
     public void runShouldBeExecuted() throws Exception {
         Map<String, Object> caseData = new HashMap<>();
-        mockTasksExecution(caseData, generalReferralDecisionDateTask);
+        mockTasksExecution(
+            caseData,
+            generalReferralDecisionDateTask,
+            generalReferralDataTask,
+            generalReferralFieldsRemovalTask);
 
         generalConsiderationWorkflow.run(CaseDetails.builder().caseData(caseData).build());
 
-        verifyTaskWasCalled(caseData, generalReferralDecisionDateTask);
+        verifyTasksCalledInOrder(
+            caseData,
+            generalReferralDecisionDateTask,
+            generalReferralDataTask,
+            generalReferralFieldsRemovalTask);
     }
 }
