+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/workflows/generalreferral/GeneralReferralWorkflowTest.java
@@ -8,14 +8,12 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralApplicationAddedDateTask;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralReferralDataTask;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.generalreferral.GeneralReferralFieldsRemovalTask;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import static uk.gov.hmcts.reform.divorce.orchestration.testutil.Verificators.mockTasksExecution;
-import static uk.gov.hmcts.reform.divorce.orchestration.testutil.Verificators.verifyTasksCalledInOrder;
+import static uk.gov.hmcts.reform.divorce.orchestration.testutil.Verificators.verifyTaskWasCalled;
 
 @RunWith(MockitoJUnitRunner.class)
 public class GeneralReferralWorkflowTest extends TestCase {
@@ -23,32 +21,16 @@ public class GeneralReferralWorkflowTest extends TestCase {
     @Mock
     private GeneralApplicationAddedDateTask generalApplicationAddedDateTask;
 
-    @Mock
-    private GeneralReferralDataTask generalReferralDataTask;
-
-    @Mock
-    private GeneralReferralFieldsRemovalTask generalReferralFieldsRemovalTask;
-
     @InjectMocks
     private GeneralReferralWorkflow generalReferralWorkflow;
 
     @Test
     public void runShouldBeExecuted() throws Exception {
         Map<String, Object> caseData = new HashMap<>();
-        mockTasksExecution(
-            caseData,
-            generalApplicationAddedDateTask,
-            generalReferralDataTask,
-            generalReferralFieldsRemovalTask
-        );
+        mockTasksExecution(caseData, generalApplicationAddedDateTask);
 
         generalReferralWorkflow.run(CaseDetails.builder().caseData(caseData).build());
 
-        verifyTasksCalledInOrder(
-            caseData,
-            generalApplicationAddedDateTask,
-            generalReferralDataTask,
-            generalReferralFieldsRemovalTask
-        );
+        verifyTaskWasCalled(caseData, generalApplicationAddedDateTask);
     }
 }
